+++ integrations/execute.js
@@ -19,14 +19,15 @@ function debounce(fn, ms) {
 
 module.exports = function $(command, options = {}) {
   let abortController = new AbortController()
-  let cwd = resolveToolRoot()
+  let root = resolveToolRoot()
+  let cwd = options.cwd ?? root
 
   let args = options.shell
     ? [command]
     : (() => {
         let args = command.split(' ')
         command = args.shift()
-        command = command === 'node' ? command : path.resolve(cwd, 'node_modules', '.bin', command)
+        command = command === 'node' ? command : path.resolve(root, 'node_modules', '.bin', command)
         return [command, args]
       })()